+++ src/callstack.cpp
@@ -1,9 +1,8 @@
 #include "callstack.h"
 
-
 void CallStack::Push(CallFrame cf)
 {
-    if(s.size() == UINT16_MAX)
+    if (s.size() == UINT8_MAX)
     {
         std::cout << "Stack overflow." << std::endl;
         exit(3);
